+++ packages/react/src/hooks/useIsEncrypted.ts
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { LocalParticipant, Participant } from 'livekit-client';
+import { LocalParticipant, Participant, Room } from 'livekit-client';
 import { encryptionStatusObservable } from '@livekit/components-core';
 import { useEnsureParticipant, useEnsureRoom } from '../context';
 import { useObservableState } from './internal';
@@ -7,9 +7,16 @@ import { useObservableState } from './internal';
 /**
  * @alpha
  */
-export function useIsEncrypted(participant?: Participant) {
+export interface UseIsEncryptedOptions {
+  room?: Room;
+}
+
+/**
+ * @alpha
+ */
+export function useIsEncrypted(participant?: Participant, options: UseIsEncryptedOptions = {}) {
   const p = useEnsureParticipant(participant);
-  const room = useEnsureRoom();
+  const room = useEnsureRoom(options.room);
 
   const observer = React.useMemo(() => encryptionStatusObservable(room, p), [room, p]);
   const isEncrypted = useObservableState(